+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Fixed
 
+- [TD-5402] Reference data info is not showed in implementation details
 - [TD-5455] Fixed Lineage selection is not displaying any domain in case that no
   lineage structure is in the data catalog